adding timeWait function, improving display's bugs, adding level notion for difficulty, other updates

+++ game/inc/gameObject.h
@@ -29,7 +29,7 @@ public:
 private:
     int m_x;
     int m_y;
-    int m_health;    
+    uint8_t m_health;    
 };
 
 #endif
